Final v1.7 commit

+++ src/utilities/cache-invalidator.js
@@ -1,52 +1,50 @@
 /**
  * Name: Semantic cache entry invalidator
- * Description: This module runs a cron job at the pre-configured schedule to invalidate (delete)
- * expired entries in the semantic cache table - apigtwycache.
+ * Description: This class implements a scheduler that periodically runs a process to invalidate (delete)
+ * expired cache entries in the semantic cache table - apigtwycache. 
  *
  * Author: Ganesh Radhakrishnan (ganrad01@gmail.com)
  * Date: 03-15-2024
  *
  * Notes:
  * ID04272024: ganrad: Centralized logging with winstonjs
- *
+ * ID05062024: ganrad: Converted this module to a class
 */
 
 const path = require('path');
 const scriptName = path.basename(__filename);
 const logger = require("./logger.js");
 
-const cacheTbl = require("../services/cp-pg.js");
+const cacheDao = require("../services/cp-pg.js");
 var cron = require('node-cron');
 
-const deleteRowStmts = [
-  "DELETE FROM apigtwycache WHERE (aiappname = $1) AND (timestamp_ < CURRENT_TIMESTAMP - INTERVAL $2)"
-  ];
+const { AzAiServices } = require("./app-gtwy-constants.js");
 
-function runCacheInvalidator(schedule, ctx) {
-  let context = ctx
+class CacheEntryInvalidator {
 
-  cron.schedule(schedule, () => {
-    // console.log(`*****\nrunCacheInvalidator(): ${new Date().toLocaleString()}`);
-    logger.log({level: "info", message: "[%s] runCacheInvalidator(): Cron schedule=[%s]", splat: [scriptName, schedule]});
+  constructor() {
+  }
 
-    let values;
-    let query;
+  async runSchedule(schedule, ctx) {
 
-    context.applications.forEach(async (app) => {
-      if ( app.cacheSettings.useCache ) {
-        let entryExpiry = app.cacheSettings.entryExpiry;
+    cron.schedule(schedule, () => {
+      // console.log(`*****\nrunCacheInvalidator(): ${new Date().toLocaleString()}`);
+      logger.log({level: "info", message: "[%s] %s.runSchedule(): Cron schedule=[%s]", splat: [scriptName, this.constructor.name, schedule]});
 
-        if ( entryExpiry ) {
-        // values = [app.appId, entryExpiry]
-        // await cacheTbl.deleteData(deleteRowStmts[0],values);
+      let query;
+      ctx.applications.forEach(async (app) => {
+        if ( (app.appType === AzAiServices.OAI) && app.cacheSettings.useCache ) {
+          let entryExpiry = app.cacheSettings.entryExpiry;
 
-          query = `DELETE FROM apigtwycache WHERE (aiappname = '${app.appId}') AND (timestamp_ < CURRENT_TIMESTAMP - INTERVAL '${entryExpiry}')`
-          // console.log(`  appId=${app.appId}\n  entryExpiry=${entryExpiry}\n  query=${query}`);
-          await cacheTbl.deleteData(query,null);
+          if ( entryExpiry ) {
+            query = `DELETE FROM apigtwycache WHERE (aiappname = '${app.appId}') AND (timestamp_ < CURRENT_TIMESTAMP - INTERVAL '${entryExpiry}')`
+            // console.log(`  appId=${app.appId}\n  entryExpiry=${entryExpiry}\n  query=${query}`);
+            await cacheDao.deleteData(query,null);
+          };
         };
-      };
+      });
     });
-  });
-}
+  } // end of runSchedule()
+} // end of class
 
-module.exports = runCacheInvalidator;
+module.exports = CacheEntryInvalidator;
